bot/requirements.txt
discord.py==2.3.2
rcon==1.2.0
python-dotenv==1.0.0